vec3 signatures

+++ main.cc
@@ -12,8 +12,8 @@ int main() {
     for (int j = height-1; j >= 0; j--) {
         std::cerr << "\r" << j << " lines left" << ' ' << std::flush;
         for (int i = 0; i < width; i++) {
-            auto r = double(i) / (width-1);
-            auto g = double(j) / (height-1);
+            auto r = float(i) / (width-1);
+            auto g = float(j) / (height-1);
             auto b = 0.25;
 
             int ir = static_cast<int>(255.999 * r);
